requirements.txt
requests
pyyaml
openpyxl
bilibili-api-python